requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-jwt-extended==4.5.3
flask-cors==4.0.0
flask-mail==0.9.1
werkzeug==2.3.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
web3==6.11.0
requests==2.31.0
bcrypt==4.0.1
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0